+++ projects/neural_network/src/optimizer/imp/sgd.rs
@@ -20,7 +20,7 @@ impl Optimizer for SGD {
     fn update<P, G>(&self, params: &mut P, grads: &G)
     where
         P: Add<G, Output = P> + Clone,
-        G: Mul<f64, Output = G> + Clone,
+        G: Mul<f32, Output = G> + Clone,
     {
         *params = params.clone() + grads.clone() * (-self.lr.value());
     }
